+++ tests/empty.R
@@ -1,7 +1,7 @@
 library(markdown)
 f = tempfile()
 if (file.create(f)) {
-   markdownToHTML(f, options = 'fragment_only')
+   markdownToHTML(f, options = '-standalone')
    markdownToHTML(f)
    unlink(f)
 }